Changed public constants to be readonly static where possible
Constant members are copied at compile time to the call sites, instead of being fetched at runtime.
https://sonarcloud.io/coding_rules?open=csharpsquid%3AS2339&rule_key=csharpsquid%3AS2339

+++ src/GameLogic/PlayerActions/Items/MoveItemAction.cs
@@ -113,7 +113,7 @@ namespace MUnique.OpenMU.GameLogic.PlayerActions.Items
                     {
                         Rows = StoreRows,
                         StartIndex = FirstStoreItemSlotIndex,
-                        EndIndex = FirstStoreItemSlotIndex + StoreSize,
+                        EndIndex = (byte)(FirstStoreItemSlotIndex + StoreSize),
                         Storage = player.ShopStorage
                     };
                     break;
